Split default implementations into separate classes
